Handle n750 hardware differences
So unlike the n600 the n750 does not actually have a hardware eth1 port
it's really on the same switch as the eth0 port and vlan's are used to
seperate them. I suppose this was a cost cutting mesure and makes me wonder
about the maximum throughput of the router vs the n600.

Regardless this is an example of why copying and pasting the network.yml.j2
isn't the best idea even on supposedly nearly identical hardware.