Get timestamps working
So to get the 'default' option in the babeld config you have to first
create an interface block with no ifname, it's kinda counterintutive.

Also in this patch we should restart networking instead of starting it
this way we apply our static interface parameters instead of having
babel generate addresses. Still not sure if that was a good idea
as a patchset.