Finish addressing fix for gateway devices
In a previous commit I removed the gateway address as the ip from the
bridge tunnel, this was correct in that using that address was confusing
but incorrect in that having no address with a greater than /128 assigned
would prevent babel routing. We now assign a much safer and less confusing
fe80 address, this prevents routing confusion while being safer since the fe80
domain is per interface.