Automated updates
This adds functionality for automated updates, we install the
package althea-cron-jobs which sets up a cron script and job
which will check for updated packages and install them every
morning at 4:30am. If an update is found it will be installed
and the system rebooted. Also added are the required packages
for https package downloads via wget from updates.altheamesh.com