Fix useless package feed
By default all feeds in feeds.conf are added to the resulting firmware
using openwrt.org links. This obviously doesn't work out so nicely for
the Althea feed, which is source only in the first place. We insert the
compiled packages feed with the firmware build. This commit just ignores
the source only feed we include for building.