Fix 5ghz mesh mode on omnia
Turns out it doesn't work at all on the Omnia, we make the odd choice
to run off of 2.4 instead for people who must have this feature.

Accross the board the network settings for wifi_mesh where all misnamed
and not setup right. I'm not totally happy with this fix for that as it
seems to leave some cruft in the generated configs but UCI seems fine with it.