Update DNS infrastructure
Nord vpn moved their server, and instead of just turning it off
they responded with redirects.

This patches that server out and adds DNS servers that I expect to
be up for the long haul. Cloudflare, Hurricane Electric, and L3

We also increased the dns cache size since the default size 150 is not
really up to date with a world where CNN.com queries 100+ domains.