Fix dev branch and rc releases to use rc package feed
So the dev profile having keys in it doesn't do much good, also
the rc builds need to use the dev profile so that they update on the
rc package feed. But I'm not sure if that really matters all that much
since the only difference between the builds at this point is the rc
package feed.