More stable DHCP configuration
Ideally we will load balance dns requests over as many servers
as possible.

This takes a moment to improve that config a good split between
primaries and secondaries from 5 different providers between the
client DNS and the router operation DNS.

Neither set is all primaries or all secondaries and I've taken the
chance to bump the cache by a factor of 4.