Add support for ea7300v2
This router has the potential to be our new low cost pick, provided
it can pass some field testing. I don't really look forward to maintaining
a mips device in the field due to the dynamic linking to glibc, but if it
works well it will probably be worth it.