Fix supported image builds for OpenWRT 21.02 development
This regenerates all supported device configs in order to provide
compatibility with 21.02, it also removes the n600 target which is now
no longer upstream. Finally it upgrades the n750 target to the ath79
backend.

This does not modify unsupported targets, which probably need attention
as well to get working again.