Add magic ip for WAN peers
This adds a magic ip address for Althea nodes that we would like to peer
to on existing internal networks.

What happens is that routers will dial out to this address and contact
the Althea node using their wan port in situations where we do have a
routed network, and a node running Althea somewhere, but no way to use
our L2 peer discovery.