Remove tplink A6 pin
We are still getting reports of issues with the tp link wifi. Although
it's not clear if they are related to driver issues or to other OpenWRT
related issues.

Given the lack of direct evidence I'm going to roll this back to the
default git pin and then we can debug an instance live, potentially
applying specific patches. I'm concerned that if we remain on master
we'll continue to think the wifi issue is unfixed due to other unrelated
problems in master that we don't see on other routers simply becuase
they are not built with the same firmware.