Fix recursive variable assignment
In the commit 'Add exit build support' device_template was disambiguated
between the exit build and client builds using a conditional assignment
in defaults. If the conditional is true there's no problem, but if it's
false we learn that Ansible does not interpret this as modifying the
existing value of the variable, but instead as a recursive condition
that is impossible to evalute.

This patch resolves the issue by renaming the variable.