Add initial support for Netgear R7800
The netgear r7800 and zyxel 2Z have the same switch chip (QCA8337) and the current swconfig driver is terrible and unable to do VLANs and prevents port toggling.  Some very nice work in the kernel is happening to make good DSA drivers in kernel 5.15.  But openwrt master is currently on 5.10, so it'll be a while ... revisit this when kernel > 5.15 lands.