Update config for exit subnets feature
This patch updates the rita config, specifcially the exit configuration
for the new subnets feature. This removes the need to constantly migrate
the config in rita. This also renames the 'exits' field to 'new_exits'
this is required as part of the migration of existing devices and will
be migrated back in the next release.

Co-authored-by: Justin Kilpatrick <justin@althea.net>