Roll back to WPA2
The use of wpa2/wpa3 mixed mode introduced in openwrt 21 has caused a
lot of problems for subsets of customer devices, laptops, printers,
smart fridges etc that can't connect to a wpa2 mixed mode network for
various reasons. This change rolls back the default while we work on a
toggle system to allow users to select their desired security level.