Add exit dns server
This adds the exit internal dns server as the primary server for both
internal and external traffic resolution.

I'm not sure how exactly this will play with gateway nodes yet, but what
should happen there is that the resolution from the internal server
fails and then succeeds when trying the next server.