Exit appliance firmware
This patch modifies the way exit firmware builds going forward. Then
Thomas introduced the rita-exit package and added config details to this
repo the exit code itself was kept mostly the same. Meaning it required
full pre-configuration of all details and unlike the normal firmware
images was not at all functional 'out of the box'.

This patch is the first step in that direction creating a exit image
that is flashable without full configuration. Extensive changes in
althea_rs, as well as the creation of an exit dashboard. Will be
required before this is usable.