Local dhcp server must be a local ip
This patch fixes a somewhat obvious problem that lan clients are nat'ed
before they reach 172.168.0.1 so therefore they would ignore this dns
server option. By changing the value to the 192.168.10.1 clients will
instead direct their requests to the router dns server.

This is paired with a larger change based on the assumption that if the
client wants us to provide a dns server it should be our own and our own
only. The stock openwrt config does not infact recomend any dhcp servers
and in some ways that's probably the way to go to avoid the issues we're
seeing with rate limited dns requests.