Update Exit firmware building
Up until now this repo has supported building exit firmwares, but exit
firmwares where not exceptionally useful by virtue of the design of the
exit software itself. Since private information such as postgresql
credentials had to be built into the image for it to be usable and no
dashboard existed to assist in field configuration exit images while
technically possible have only ever been built once or twice.

This patch updates the firmware builder to reflect upstream changes in
beta 22 of Rita Exit which make it entierly feasible to build out ship
and forget exit firmwares that don't require any special configuration
to be used in the field.