Modify reset button behavior
The default OpenWRT behavior for a long press of the reset button over 5 secs is to completely erase the changed files overlay on the squashfs filesystem.  This commit replaces that to only delete the rita dashboard password and to remove only the wifi config from the overlay directory.  It does not do anything regarding the SSH password.