Update /etc/config/dhcp
After fighting with this some on my home setup I realized the servers
list is not specifying what servers to suggest to the client but instead
what servers dnsmasq itself is using. The self-referential nature of
this configuration error means it takes an inordinate amount of time for
things to 'start working'.